ELT pipeline broken into producer/consumer queues (non-blocking)
