+++ src/dumpty/main.py
@@ -3,11 +3,14 @@ import json
 import logging
 import os
 import sys
+import time
 from concurrent.futures import Future, ThreadPoolExecutor, as_completed
 from datetime import datetime
 from math import ceil
 from pathlib import Path
-from typing import List
+from queue import Queue, Empty
+from threading import Thread
+from typing import Callable, List
 
 import psutil
 from alive_progress import alive_bar
@@ -15,14 +18,15 @@ from jinja2 import Environment, FileSystemLoader, Template
 from sqlalchemy import create_engine
 from tenacity import (Retrying, stop_after_attempt, stop_after_delay,
                       wait_random_exponential)
+from tqdm import tqdm
 
 from dumpty import logger
 from dumpty.config import Config
-from dumpty.gcp import (bigquery_create_dataset, bigquery_load,
-                        upload_from_string, get_size_bytes)
-from dumpty.introspector import Extract, Introspector
-from dumpty.spark import LocalSpark
-from dumpty.util import filter_shuffle, normalize_str
+from dumpty.extract import Extract
+from dumpty.extract_db import ExtractDB
+from dumpty.gcp import bigquery_create_dataset
+from dumpty.pipeline import Pipeline
+from dumpty.util import filter_shuffle
 
 
 def config_from_args(argv) -> Config:
@@ -112,62 +116,9 @@ def config_from_args(argv) -> Config:
     return config
 
 
-def extract_and_load(table_name: str, config: Config, spark: LocalSpark, introspector: Introspector, retryer: Retrying):
-    # Introspect if needed
-    extract = retryer(introspector.introspect, table_name,
-                      config.partitioning_threshold)
-
-    # Extract the table with Spark
-    extract.extract_uri = None
-    if config.target_uri is not None:
-        extract_uri = retryer(spark.full_extract, extract, config.target_uri)
-        extract.extract_uri = extract_uri
-        extract.extract_date = datetime.now()
-
-        # Suggest a recommended partition size based on the actual extract size (for next run)
-        if extract.partitions > 0 and "gs://" in extract_uri:  # only resizes based on GCS targets, for now
-            extract.gcs_bytes = get_size_bytes(extract_uri)
-            recommendation = ceil(
-                extract.gcs_bytes / config.target_partition_size_bytes)
-            if recommendation != extract.partitions:
-                logger.info(
-                    f"Adjusted partitions on {extract.name} from {extract.partitions} to {recommendation} for next run")
-                extract.partitions = recommendation
-                extract.introspect_date = None  # triggers new introspection next run
-
-        # Get JSON-formatted table schema
-        json_schema = json.dumps(extract.bq_schema, indent=4)
-        if "gs://" not in config.target_uri:
-            with open(f"{config.target_uri}/{normalize_str(extract.name)}/schema.json", "wt") as f:
-                f.write(json_schema)
-        else:
-            retryer(upload_from_string, json_schema,
-                    f"{config.target_uri}/{normalize_str(extract.name)}/schema.json")
-
-        # Load into BigQuery
-        if config.target_dataset is not None:
-            # Load from GCS into BQ
-            normalized_table_name = normalize_str(extract.name)
-            bq_rows: int = 0
-            bq_bytes: int = 0
-            if config.target_dataset is not None:
-                logger.info(
-                    f"Loading {extract.name} into BigQuery as {normalized_table_name} from {extract.extract_uri}")
-                bq_rows, bq_bytes = retryer(bigquery_load, extract_uri, f"{config.target_dataset}.{normalized_table_name}",
-                                            config.spark.format, extract.bq_schema, "Loaded by Dumpty")
-            extract.rows_loaded = bq_rows
-            extract.bq_bytes = bq_bytes
-
-    return extract
-
-
 def main(args=None):
 
     config: Config = config_from_args(args)
-
-    # Initialize SqlAlchemy
-    engine = create_engine(config.sqlalchemy.url, pool_size=config.sqlalchemy.pool_size, connect_args=config.sqlalchemy.connect_args,
-                           max_overflow=config.sqlalchemy.max_overflow, pool_pre_ping=True, echo=False)
 
     # Default retry for network operations: 2^x * 1 second between each retry, starting with 5s, up to 60s, die after 5 minutes of retries
     # reraise=True places the exception at the END of the stack-trace dump
@@ -191,65 +142,83 @@ def main(args=None):
         "warnings": []
     }
 
-    extract_jobs: List[Future] = []
-    completed: List[Extract] = []
-    with Introspector(config.tinydb_database_file, engine, config.schema) as introspector:
+    # Initialize SqlAlchemy
+    engine = create_engine(config.sqlalchemy.url, pool_size=config.sqlalchemy.pool_size, connect_args=config.sqlalchemy.connect_args,
+                           max_overflow=config.sqlalchemy.max_overflow, pool_pre_ping=True, echo=False)
 
-        if config.reconcile:
-            # Check if tables being requested actually exist in SQL database before doing anything else
-            # This can be very slow for databases with thousands of tables so it is off by default
-            introspector.reconcile(config.tables)
+    with ExtractDB(config.tinydb_database_file, default_table_name=config.schema) as extract_db:
+        with Pipeline(engine, retryer, config) as pipeline:
 
-        with LocalSpark(config, config.target_uri) as spark:
+            if config.reconcile:
+                # Check if tables being requested actually exist in SQL database before doing anything else
+                # This can be very slow for databases with thousands of tables so it is off by default
+                pipeline.reconcile(config.tables)
 
-            # Throttles the number of jobs submitted to Spark (and # concurrent pysql DB connections)
-            with ThreadPoolExecutor(max_workers=config.job_threads) as executor:
+            for table in config.tables:
+                extract: Extract = extract_db.get(table)
+                pipeline.submit(extract)
 
-                for table in config.tables:
-                    extract_jobs.append(executor.submit(
-                        extract_and_load, table, config, spark, introspector, retryer))
-
-                with alive_bar(len(extract_jobs), dual_line=True, stats=False, disable=not config.progress_bar) as bar:
-                    for future in as_completed(extract_jobs):
-                        try:
-                            extracted_table: Extract = future.result()
-                            introspector.save(extracted_table)
-                            completed.append(extracted_table)
-                            logger.info(f"{extracted_table.name} complete")
-                        except Exception as ex:
-                            logger.error(ex)
-                            executor.shutdown(
-                                wait=False, cancel_futures=True)
-                            raise ex
-                        summary['tables'].append(extracted_table.name)
-                        if config.target_dataset is not None:
-                            if not extracted_table.consistent():
-                                warning = f"{extracted_table.name}: row count mismatch (expected: {extracted_table.rows}, loaded: {extracted_table.rows_loaded}+"
-                                logger.warning(warning)
-                                summary['warnings'].append(warning)
-                        bar.text = f"| {datetime.now().strftime('%H:%M:%S')} | System CPU: {psutil.cpu_percent()}% | Memory: {psutil.virtual_memory()[2]}%"
+            count = 0
+            with alive_bar(len(config.tables), dual_line=True, stats=False, disable=not config.progress_bar) as bar:
+                while count < len(config.tables):
+                    try:
+                        extract = pipeline.done_queue.get(timeout=1)
+                        extract_db.save(extract)
+                        count += 1
                         bar()
+                    except Empty:
+                        pass
+                    finally:
+                        bar.text = f"introspecting: [{pipeline.introspect_queue.unfinished_tasks}] extracting: [{pipeline.extract_queue.unfinished_tasks}] loading: [{pipeline.load_queue.unfinished_tasks}]"
 
-                logger.info("Extraction complete, shutting down")
+    #         # Throttles the number of jobs submitted to Spark (and # concurrent pysql DB connections)
+    #         with ThreadPoolExecutor(max_workers=config.job_threads) as executor:
 
-    # Summarize
-    summary['end_date'] = datetime.now()
-    summary['elapsed_s'] = round(
-        (summary['end_date'] - summary['start_date']).total_seconds())
+    #             for table in config.tables:
+    #                 extract_jobs.append(executor.submit(
+    #                     extract_and_load, table, config, spark, introspector, retryer))
 
-    if config.target_dataset is not None:
-        summary['consistent'] = all(x.consistent() for x in completed)
-        summary['bq_bytes'] = sum(x.bq_bytes for x in completed)
+    #             with alive_bar(len(extract_jobs), dual_line=True, stats=False, disable=not config.progress_bar) as bar:
+    #                 for future in as_completed(extract_jobs):
+    #                     try:
+    #                         extracted_table: Extract = future.result()
+    #                         introspector.save(extracted_table)
+    #                         completed.append(extracted_table)
+    #                         logger.info(f"{extracted_table.name} complete")
+    #                     except Exception as ex:
+    #                         logger.error(ex)
+    #                         executor.shutdown(
+    #                             wait=False, cancel_futures=True)
+    #                         raise ex
+    #                     summary['tables'].append(extracted_table.name)
+    #                     if config.target_dataset is not None:
+    #                         if not extracted_table.consistent():
+    #                             warning = f"{extracted_table.name}: row count mismatch (expected: {extracted_table.rows}, loaded: {extracted_table.rows_loaded}+"
+    #                             logger.warning(warning)
+    #                             summary['warnings'].append(warning)
+    #                     bar.text = f"| {datetime.now().strftime('%H:%M:%S')} | System CPU: {psutil.cpu_percent()}% | Memory: {psutil.virtual_memory()[2]}%"
+    #                     bar()
 
-    if config.target_uri is not None:
-        summary['gcs_bytes'] = sum(x.gcs_bytes for x in completed)
+    #             logger.info("Extraction complete, shutting down")
 
-    with open(config.log_file, "w") as outfile:
-        outfile.write(json.dumps(summary, indent=4, default=str))
+    # # Summarize
+    # summary['end_date'] = datetime.now()
+    # summary['elapsed_s'] = round(
+    #     (summary['end_date'] - summary['start_date']).total_seconds())
 
-    if not len(summary['warnings']) == 0:
-        logger.warning(
-            f"{len(summary['tables'])} tables loaded, with warnings")
+    # if config.target_dataset is not None:
+    #     summary['consistent'] = all(x.consistent() for x in completed)
+    #     summary['bq_bytes'] = sum(x.bq_bytes for x in completed)
+
+    # if config.target_uri is not None:
+    #     summary['gcs_bytes'] = sum(x.gcs_bytes for x in completed)
+
+    # with open(config.log_file, "w") as outfile:
+    #     outfile.write(json.dumps(summary, indent=4, default=str))
+
+    # if not len(summary['warnings']) == 0:
+    #     logger.warning(
+    #         f"{len(summary['tables'])} tables loaded, with warnings")
 
 
 if __name__ == '__main__':
