+++ src/dumpty/sql.py
@@ -1,54 +1,52 @@
-from abc import ABC, abstractmethod
-
-from sqlalchemy.engine import Engine
+from sqlalchemy import Column, Table
 from sqlalchemy.sql.sqltypes import *
 
+from dumpty import logger
+from dumpty.util import normalize_str
 
-class Sql(ABC):
-    """Class for handling dialect-specific SQL statements that can't be executed using SqlAlchemy core (query builder)
-    Currently, there are none of these, but this class is being left here for future use...
+
+def bq_schema(table: Table) -> list[dict]:
     """
+    Returns Table schema in BigQuery JSON schema format
+    """
+    col: Column
+    schemas = []
+    for col in table.columns:
+        schema = {}
+        schema['name'] = normalize_str(col.name)
+        schema['mode'] = "Nullable" if col.nullable else "Required"
 
-    def __init__(self, schema: str, engine: Engine):
-        self.engine = engine
-        self.schema = schema
-        super().__init__()
+        if isinstance(col.type, (SmallInteger, Integer, BigInteger)):
+            schema['type'] = "INT64"
+        elif isinstance(col.type, (DateTime)):
+            schema['type'] = "DATETIME"
+        elif isinstance(col.type, (Date)):
+            schema['type'] = "DATE"
+        elif isinstance(col.type, (Float, REAL)):
+            schema['type'] = "FLOAT64"
+        elif isinstance(col.type, (String)):
+            schema['type'] = "STRING"
+        elif isinstance(col.type, (Boolean)):
+            schema['type'] = "BOOL"
+        elif isinstance(col.type, (LargeBinary)):
+            schema['type'] = "BYTES"
+        elif isinstance(col.type, (Numeric)):
+            p = col.type.precision
+            s = col.type.scale
+            if s == 0 and p <= 18:
+                schema['type'] = "INT64"
+            elif (s >= 0 and s <= 9) and ((max(s, 1) <= p) and p <= s+29):
+                schema['type'] = "NUMERIC"
+                schema['precision'] = p
+                schema['scale'] = s
+            elif (s >= 0 and s <= 38) and ((max(s, 1) <= p) and p <= s+38):
+                schema['type'] = "BIGNUMERIC"
+                schema['precision'] = p
+                schema['scale'] = s
+        else:
+            logger.warning(
+                f"Unmapped type in {table.name}.{col.name} ({col.type}), defaulting to STRING")
+            schema['type'] = "STRING"
 
-    @abstractmethod
-    def get_row_count(self, table: str) -> int:
-        """
-        Returns number of rows in table, and number of seconds to get that count
-        """
-        pass
-
-
-def factory(schema: str, engine: Engine) -> Sql:
-    if engine.dialect.name == "mssql":
-        return Mssql(schema, engine)
-    elif engine.dialect.name == "oracle":
-        return Oracle(schema, engine)
-    else:
-        raise Exception(
-            f"Unsupported dialect {engine.dialect.name}")
-
-
-class Mssql(Sql):
-    def __init__(self, schema: str, engine: Engine):
-        super().__init__(schema, engine)
-
-    def get_row_count(self, table: str) -> int:
-        sql = f"SELECT COUNT(*) AS count FROM [{self.schema}].[{table}]"
-        with self.engine.connect() as connection:
-            result = connection.execute(sql).scalar()
-        return int(result)
-
-
-class Oracle(Sql):
-    def __init__(self, schema: str, engine: Engine):
-        super().__init__(schema, engine)
-
-    def get_row_count(self, table: str) -> int:
-        sql = f"SELECT COUNT(*) AS count FROM {self.schema}.{table}"
-        with self.engine.connect() as connection:
-            result = connection.execute(sql).scalar()
-        return int(result)
+        schemas.append(schema)
+    return schemas
